pyproject.toml
[project]
name = "llm-for-gitlab-code-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "openai>=1.97.0",
    "pydantic-settings>=2.10.1",
    "python-gitlab>=6.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
]
